# apps/wallet-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "relational-rust-server",
    "description": "Custodial Avalanche wallet service running inside Intel SGX enclave with Gramine.",
    "contact": {
      "name": "Relational Network"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/recurring/payment": {
      "post": {
        "tags": [
          "Recurring"
        ],
        "operationId": "create_recurring_payment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurringPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/v1/recurring/payment/{recurring_payment_id}": {
      "put": {
        "tags": [
          "Recurring"
        ],
        "operationId": "update_recurring_payment",
        "parameters": [
          {
            "name": "recurring_payment_id",
            "in": "path",
            "description": "Identifier of the recurring payment to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecurringPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Recurring"
        ],
        "operationId": "delete_recurring_payment",
        "parameters": [
          {
            "name": "recurring_payment_id",
            "in": "path",
            "description": "Identifier of the recurring payment to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/v1/recurring/payment/{recurring_payment_id}/last-paid-date": {
      "put": {
        "tags": [
          "Recurring"
        ],
        "operationId": "update_last_paid_date",
        "parameters": [
          {
            "name": "recurring_payment_id",
            "in": "path",
            "description": "Identifier of the recurring payment to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLastPaidDateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v1/recurring/payments": {
      "get": {
        "tags": [
          "Recurring"
        ],
        "operationId": "list_recurring_payments",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/WalletAddress"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecurringPayment"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRecurringPaymentRequest": {
        "type": "object",
        "description": "Request to create a recurring payment schedule.",
        "required": [
          "wallet_id",
          "wallet_public_key",
          "recipient",
          "amount",
          "currency_code",
          "payment_start_date",
          "frequency",
          "payment_end_date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Payment amount."
          },
          "currency_code": {
            "type": "string",
            "description": "Currency code (e.g., \"AVAX\", \"USDC\")."
          },
          "frequency": {
            "type": "integer",
            "format": "int32",
            "description": "Frequency in days between payments."
          },
          "payment_end_date": {
            "type": "integer",
            "format": "int32",
            "description": "End date (Unix timestamp in days, 0 = no end)."
          },
          "payment_start_date": {
            "type": "integer",
            "format": "int32",
            "description": "Start date (Unix timestamp in days)."
          },
          "recipient": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "Recipient address for payments."
          },
          "wallet_id": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The source wallet (must be owned by the user)."
          },
          "wallet_public_key": {
            "type": "string",
            "description": "Public key for the wallet."
          }
        }
      },
...
      "RecurringPayment": {
        "type": "object",
        "description": "A scheduled recurring payment configuration.\n\nRecurring payments allow automatic transfers on a schedule. The actual\nexecution logic is handled by a separate service.",
        "required": [
          "id",
          "wallet_id",
          "wallet_public_key",
          "recipient",
          "amount",
          "currency_code",
          "payment_start_date",
          "frequency",
          "payment_end_date",
          "last_paid_date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Payment amount."
          },
          "currency_code": {
            "type": "string",
            "description": "Currency code (e.g., \"AVAX\", \"USDC\")."
          },
          "frequency": {
            "type": "integer",
            "format": "int32",
            "description": "Frequency in days between payments."
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for this payment schedule."
          },
          "last_paid_date": {
            "type": "integer",
            "format": "int32",
            "description": "Last payment date (Unix timestamp in days)."
          },
          "payment_end_date": {
            "type": "integer",
            "format": "int32",
            "description": "End date (Unix timestamp in days, 0 = no end)."
          },
          "payment_start_date": {
            "type": "integer",
            "format": "int32",
            "description": "Start date (Unix timestamp in days)."
          },
          "recipient": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The recipient address for payments."
          },
          "wallet_id": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "The source wallet for payments."
          },
          "wallet_public_key": {
            "type": "string",
            "description": "Public key of the wallet (for verification)."
          }
        }
      },
...
      "UpdateLastPaidDateRequest": {
        "type": "object",
        "description": "Request to update the last paid date for a recurring payment.",
        "required": [
          "recurring_payment_id",
          "last_paid_date"
        ],
        "properties": {
          "last_paid_date": {
            "type": "integer",
            "format": "int32",
            "description": "New last paid date (Unix timestamp in days)."
          },
          "recurring_payment_id": {
            "type": "string",
            "description": "ID of the recurring payment."
          }
        }
      },
      "UpdateRecurringPaymentRequest": {
        "type": "object",
        "description": "Request to update an existing recurring payment.",
        "required": [
          "recurring_payment_id",
          "wallet_id",
          "wallet_public_key",
          "recipient",
          "amount",
          "currency_code",
          "payment_start_date",
          "frequency",
          "payment_end_date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double",
            "description": "Updated payment amount."
          },
          "currency_code": {
            "type": "string",
            "description": "Updated currency code."
          },
          "frequency": {
            "type": "integer",
            "format": "int32",
            "description": "Updated frequency."
          },
          "payment_end_date": {
            "type": "integer",
            "format": "int32",
            "description": "Updated end date."
          },
          "payment_start_date": {
            "type": "integer",
            "format": "int32",
            "description": "Updated start date."
          },
          "recipient": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "Updated recipient address."
          },
          "recurring_payment_id": {
            "type": "string",
            "description": "ID of the recurring payment to update."
          },
          "wallet_id": {
            "$ref": "#/components/schemas/WalletAddress",
            "description": "Updated source wallet."
          },
          "wallet_public_key": {
            "type": "string",
            "description": "Updated public key."
          }
        }
      },
...
      "WalletAddress": {
        "type": "string",
        "description": "Ethereum-compatible wallet address wrapper.\n\nProvides type safety for wallet addresses throughout the API.\nFormat: `0x` followed by 40 hexadecimal characters (20 bytes).\n\n# Example\n\n```rust,ignore\nlet addr = WalletAddress::from(\"0x742d35Cc6634C0532925a3b844Bc9e7595f4aB12\");\n```"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk JWT token"
      }
    }
  },
...
}